requirements.txt
flask
python-dotenv
langchain
langchain-community
langchain-groq
langchain-text-splitters
sentence-transformers
faiss-cpu
pypdf